Collapse per-agent hook scripts into one universal secrets_guard.py
One stdlib-only script now serves Claude Code, Codex, and Cursor by
detecting the caller's dialect from the stdin payload: Cursor is
identified by its hook_event_name values (beforeReadFile /
beforeShellExecution), everything else dispatches on tool_name and
replies in the shared Claude/Codex PreToolUse shape.

Single-file with no sibling imports so the dlthub-init scaffolder can
bundle and copy it as-is; per-agent configs all point at the same file.
Codex config targets .codex/hooks.json rather than config.toml [hooks]
because of openai/codex#17532.
